% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blood_meridian}
\alias{blood_meridian}
\title{Locations from the novel Blood Meridian: or, The Evening Redness in the West by Cormac McCarthy.}
\format{
A tibble with 19 rows and 4 columns:
\describe{
\item{   location}{Place name.}
\item{   description}{Description of location.}
\item{   lng}{Longitude.}
\item{   lat}{Latitude.}
}
}
\source{
'@VanDiemen_'
}
\usage{
blood_meridian
}
\description{
Data and decriptions from the wonderful Twitter thread \url{https://twitter.com/VanDiemen_/status/1787491589948076444}
}
\keyword{datasets}
