\name{gdata-package}
\alias{gdata-package}
\alias{gdata}
\docType{package}
\title{Various R Programming Tools for Data Manipulation}
\description{
  \pkg{gdata} package provides various \R programming tools for data
  manipulation.
}
\details{
  The following are sources of information on \pkg{gdata} package:
  \tabular{ll}{
    DESCRIPTION file  \tab \code{library(help="gdata")}\cr
                      \tab \cr
    This file         \tab \code{?gdata}\cr
                      \tab \cr
    Vignette          \tab \code{vignette("unknown")}\cr
                      \tab \code{vignette("mapLevels")}\cr
                      \tab \cr
    Some help files   \tab \code{\link{Args}}\cr
                      \tab \code{\link{ll}}\cr
                      \tab \code{\link{write.fwf}}\cr
                      \tab \cr
    News              \tab \code{news(package="gdata")}\cr
                      \tab \cr
    List of functions \tab \code{help(package="gdata")}\cr
  }
}
\keyword{package}
