% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdata-package.R
\docType{package}
\name{gdata-package}
\alias{gdata}
\alias{gdata-package}
\title{gdata: Data for R class and homework.}
\description{
Data for R class and homework.
}
\author{
\strong{Maintainer}: Andrew Griffen \email{andrew.griffen@gmail.com}

}
\keyword{internal}
