% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdata-package.R
\docType{package}
\name{gdata-package}
\alias{gdata}
\alias{gdata-package}
\title{gdata: Datasets for R for Empirical Economics Research}
\description{
Example datasets used for my course "R for Empirical Economics Research".
}
\author{
\strong{Maintainer}: Andrew Griffen \email{andrew.griffen@gmail.com}

}
\keyword{internal}
