
\name{gdata-package}
\alias{gdata-package}
\alias{gdata}

\docType{package}

\title{Various R programming tools for data manipulation}

\description{

\pkg{gdata} package provides various \R programming tools for data
manipulation.

}

\details{
The following are sources of information on \pkg{gdata} package:
\tabular{ll}{
  DESCRIPTION file\tab \code{library(help="gdata")}\cr
  \tab \cr
  This file\tab        \code{package?gdata}\cr
  \tab \cr
  Vignette\tab         \code{vignette("unknown")}\cr
  \tab                 \code{vignette("mapLevels")}\cr
  \tab \cr
  Some help files\tab  \code{\link{read.xls}}\cr
  \tab                 \code{\link{write.fwf}}\cr
  \tab \cr
  News\tab             \code{file.show(system.file("NEWS", package="gdata"))}\cr
}

}

\keyword{package}